appease a compiler warning with wider huff_sym_t

+++ src/chuff_decode.c
@@ -4,7 +4,7 @@ uint8_t
 chuff_decode_byte(struct chuff_decode_context *ctx, const uint8_t *tbl,
                   const uint16_t *tblidx)
 {
-        ctx->chuff_ctx =
-                huff_decode_sym(&ctx->hctx, &tbl[tblidx[ctx->chuff_ctx]]);
+        ctx->chuff_ctx = (uint8_t)huff_decode_sym(
+                &ctx->hctx, &tbl[tblidx[ctx->chuff_ctx]]);
         return ctx->chuff_ctx;
 }
